requirements.txt
asgiref==3.5.2
async-timeout==4.0.3
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1
django-allauth==0.51.0
django-appconf==1.0.5
django-cleanup==8.1.0
django-config-models==2.3.0
django-constance==2.9.1
django-cors-headers==3.13.0
django-crum==0.7.9
django-environ==0.9.0
django-filter==22.1
django-imagekit==4.1.0
django-json-widget==2.0.1
django-model-utils==4.2.0
django-nested-admin==4.0.2
django-phonenumber-field==8.0.0
django-picklefield==3.1
django-redis==5.4.0
django-rest-swagger==2.2.0
django-smart-selects==1.6.0
django-waffle==3.0.0
django-wkhtmltopdf==3.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-extra-fields==3.4.1
drf-yasg==1.21.7
edx-django-utils==5.2.0
filetype==1.2.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
newrelic==8.4.0
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
pathspec==0.9.0
pbr==5.11.0
phonenumberslite==8.13.40
pilkit==2.0
Pillow==9.2.0
platformdirs==2.5.2
psutil==5.9.4
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.4.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0.1
redis==5.0.7
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.2
stevedore==4.1.0
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==1.26.11
